+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Zo.cancel_all_orders{,_ix}`.
 - BREAKING: Add heimdall to `Zo.withdraw` accounts list.
 - BREAKING: Change default commitment from `confirmed` to `finalized`.
+- BREAKING: Remove deprecated `PositionInfo.value`.
+- BREAKING: Remove deprecated `MarketInfo.funding_rate`.
 
 ## [0.1.4] - 2022-08-02